+++ app/frontend/src/javascript/controllers/admin/invoices.js
@@ -61,6 +61,14 @@ Application.Controllers.controller('InvoicesController', ['$scope', '$state', 'I
       templateUrl: '/admin/invoices/settings/editPrefix.html'
     };
 
+    // Payment Schedule PDF filename settings (and example)
+    $scope.scheduleFile = {
+      prefix: settings.payment_schedule_prefix,
+      nextId: 11,
+      date: moment().format('DDMMYYYY'),
+      templateUrl: '/admin/invoices/settings/editSchedulePrefix.html'
+    };
+
     // Invoices parameters
     $scope.invoice = {
       logo: null,
@@ -518,6 +526,38 @@ Application.Controllers.controller('InvoicesController', ['$scope', '$state', 'I
       return modalInstance.result.then(function (model) {
         Setting.update({ name: 'invoice_prefix' }, { value: model }, function (data) {
           $scope.file.prefix = model;
+          return growl.success(_t('app.admin.invoices.prefix_successfully_saved'));
+        }
+        , function (error) {
+          if (error.status === 304) return;
+
+          growl.error(_t('app.admin.invoices.an_error_occurred_while_saving_the_prefix'));
+          console.error(error);
+        });
+      });
+    };
+
+    /**
+     * Open a modal dialog allowing the user to edit the prefix of the payment schedules file names
+     */
+    $scope.openEditSchedulePrefix = function () {
+      const modalInstance = $uibModal.open({
+        animation: true,
+        templateUrl: $scope.scheduleFile.templateUrl,
+        size: 'lg',
+        resolve: {
+          model () { return $scope.scheduleFile.prefix; }
+        },
+        controller: ['$scope', '$uibModalInstance', 'model', function ($scope, $uibModalInstance, model) {
+          $scope.model = model;
+          $scope.ok = function () { $uibModalInstance.close($scope.model); };
+          $scope.cancel = function () { $uibModalInstance.dismiss('cancel'); };
+        }]
+      });
+
+      modalInstance.result.then(function (model) {
+        Setting.update({ name: 'payment_schedule_prefix' }, { value: model }, function (data) {
+          $scope.scheduleFile.prefix = model;
           return growl.success(_t('app.admin.invoices.prefix_successfully_saved'));
         }
         , function (error) {